feat(#1): dockerfile modified to run main with two controllers
