+++ intelligent-service/Dockerfile
@@ -10,7 +10,6 @@ RUN pip install pipenv
 RUN pipenv install --deploy --system --ignore-pipfile
 #RUN pipenv shell
 EXPOSE 5000
-ENV FLASK_APP=./src/app.py
-#CMD python ./src/main.py
-CMD ["flask", "run", "-h", "0.0.0.0", "-p", "5000"]
+
+CMD python ./src/main.py
